cleaning files
